requirements.txt
langchain
pydantic
python-dotenv
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu